Objects: Update movement
Make movement be constant at speed=2.
Update direction change to use unit circle.

Add global current direction which is a unit circle degrees value for which direction
the snake is moving.

Use this global current direction to update the built-in direction variable.

Set the snake default speed to 2 so that it is always moving in the direction of
the global current direction.

Remove old movement 'wasd' and 'arrow-keys' in favor of 'wd' and 'left-right arrow-keys'
so that it only checks one keypress instead of how fast your I/O (input/output) keyboard
driver operation is.

Signed-off-by: zdworkin <zachary.dworkin@gmail.com>